backend/requirements.txt
openai
huggingface_hub
requests
python-dotenv
pydantic
fastapi 
uvicorn[standard] 
sse-starlette